+++ SuperBuild/External_ROOT.cmake
@@ -1,8 +1,6 @@
 #========================================================================
-# Author: Benjamin A Thomas
-# Author: Edoardo Pasca
-# Copyright 2017, 2020, 2022 University College London
-# Copyright 2017, 2020 STFC
+# Author: Kris Thielemans
+# Copyright 2017, 2020, 2021 University College London
 #
 # This file is part of the CCP SyneRBI (formerly PETMR) Synergistic Image Reconstruction Framework (SIRF) SuperBuild.
 #
@@ -21,11 +19,10 @@
 #=========================================================================
 
 #This needs to be unique globally
-set(proj GTest)
+set(proj ROOT)
 
 # Set dependency list
-set(${proj}_DEPENDENCIES "")
-
+#set(${proj}_DEPENDENCIES "")
 
 # Include dependent projects if any
 ExternalProject_Include_Dependencies(${proj} DEPENDS_VAR ${proj}_DEPENDENCIES)
@@ -35,48 +32,50 @@ set(externalProjName ${proj})
 
 SetCanonicalDirectoryNames(${proj})
 
+
 if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalProjName}}" ) )
   message(STATUS "${__indent}Adding project ${proj}")
   SetGitTagAndRepo("${proj}")
 
   ### --- Project specific additions here
 
-  set(CMAKE_LIBRARY_PATH ${CMAKE_LIBRARY_PATH} ${SUPERBUILD_INSTALL_DIR})
-  set(CMAKE_INCLUDE_PATH ${CMAKE_INCLUDE_PATH} ${SUPERBUILD_INSTALL_DIR})
+  option(ROOT_opengl "Build ROOT with opengl" OFF)
+  option(ROOT_x11 "Build ROOT with x11" OFF)
+  mark_as_advanced(ROOT_opengl ROOT_x11)
+
+  set(${proj}_CMAKE_ARGS
+      -DPYTHON_EXECUTABLE:FILE=${PYTHON_EXECUTABLE}
+      -Dopengl=${ROOT_opengl}
+      -Dx11=${ROOT_x11}
+      -DOPENMP_INCLUDES:PATH=${OPENMP_INCLUDES}
+      -DOPENMP_LIBRARIES:PATH=${OPENMP_LIBRARIES}
+  #   -DCMAKE_INSTALL_PREFIX=${${proj}_INSTALL_DIR}
+  )
 
   ExternalProject_Add(${proj}
     ${${proj}_EP_ARGS}
     ${${proj}_EP_ARGS_GIT}
     ${${proj}_EP_ARGS_DIRS}
 
-    CMAKE_ARGS
-      -DCMAKE_PREFIX_PATH:PATH=${SUPERBUILD_INSTALL_DIR}
-      -DCMAKE_INSTALL_PREFIX:PATH=${${proj}_INSTALL_DIR}
+    CMAKE_ARGS ${${proj}_CMAKE_ARGS}
+    ${${proj}_EXTRA_CMAKE_ARGS}
     DEPENDS
         ${${proj}_DEPENDENCIES}
   )
 
-  if (${CMAKE_VERSION} VERSION_LESS "3.20")
-    # older CMake versions ignore GTest_DIR
-    set(GTest_CMAKE_ARGS
-      -DGTEST_ROOT:PATH="${GTest_INSTALL_DIR}")
-  else()
-    # We're using GTest 1.11.0 or later which exports CMake files
-    set(GTest_CMAKE_ARGS
-      -DGTest_DIR:PATH="${GTest_INSTALL_DIR}/lib/cmake/GTest")
+   else()
+      if(${USE_SYSTEM_${externalProjName}})
+        find_package(${proj} ${${externalProjName}_REQUIRED_VERSION} REQUIRED)
+        message(STATUS "USING the system ${externalProjName}, set ${externalProjName}_DIR=${${externalProjName}_DIR}")
+   endif()
+    ExternalProject_Add_Empty(${proj} DEPENDS "${${proj}_DEPENDENCIES}"
+    ${${proj}_EP_ARGS_DIRS}
+   )
   endif()
 
-else()
-      if(${USE_SYSTEM_${externalProjName}})
-        message(STATUS "USING the system ${externalProjName}, set GTest_DIR (or GTEST_ROOT for GTest < 1.8.1) if needed.")
-        find_package(${proj} ${${externalProjName}_REQUIRED_VERSION} REQUIRED)
-    endif()
-    ExternalProject_Add_Empty(${proj} DEPENDS "${${proj}_DEPENDENCIES}"
-      ${${proj}_EP_ARGS_DIRS})
-
-    if (DEFINED GTest_DIR)
-      set(GTest_CMAKE_ARGS -DGTest_DIR:PATH="${GTest_DIR}")
-    elseif (DEFINED GTEST_ROOT)
-      set(GTest_CMAKE_ARGS -DGTEST_ROOT:PATH="${GTEST_ROOT}")
-    endif()
-endif()
+  mark_as_superbuild(
+    VARS
+      ${externalProjName}_DIR:PATH
+    LABELS
+      "FIND_PACKAGE"
+  )