+++ VirtualBox/scripts/bootstrap_ubuntu.sh
@@ -1,4 +1,29 @@
 #!/bin/bash    
+#
+# Script to create the SyneRBI VM. Do not use for your own system!
+#
+# Authors: Kris Thielemans, Evgueni Ovtchinnikov, Edoardo Pasca,
+# Casper da Costa-Luis
+# Copyright 2016-2022 University College London
+# Copyright 2016-2022 Rutherford Appleton Laboratory STFC
+#
+# This is software developed for the Collaborative Computational
+# Project in Synergistic Reconstruction for Biomedical Imaging (formerly PETMR)
+# (http://www.ccpsynerbi.ac.uk/).
+
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0.txt
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+#=========================================================================
 
 set -x
 SIRFUSERNAME=sirfuser
@@ -61,22 +86,13 @@ if [ ! -d $userHOME/devel ]; then
 fi
 cd $userHOME/devel
 
-if [ ! -d $userHOME/devel/SyneRBI_VM ]; then
-  git clone https://github.com/SyneRBI/SyneRBI_VM.git
-  cd SyneRBI_VM
+if [ ! -d $userHOME/devel/SIRF-SuperBuild ]; then
+  git clone https://github.com/SyneRBI/SIRF-SuperBuild.git
+  cd SIRF-SuperBuild
 else
-  cd SyneRBI_VM
-  git pull
+  cd SIRF-SuperBuild
+  # git pull
 fi
 
-
-
-bash $userHOME/devel/SyneRBI_VM/scripts/INSTALL_prerequisites_with_apt-get.sh
-bash $userHOME/devel/SyneRBI_VM/scripts/INSTALL_CMake.sh /usr/local
-bash $userHOME/devel/SyneRBI_VM/scripts/INSTALL_python_packages.sh
-
-# port 8888 is forwarded to 8888 so the user can use the host browser
-# Therefore no browser is installed on the VM
-
 chown -R $SIRFUSERNAME:users $userHOME
-sudo -u $SIRFUSERNAME -H bash $userHOME/devel/SyneRBI_VM/scripts/UPDATE.sh
+sudo -u $SIRFUSERNAME -H bash $userHOME/devel/SIRF-SuperBuild/VirtualBox/scripts/UPDATE.sh -s