+++ docker/build_essential-ubuntu.sh
@@ -8,8 +8,8 @@ apt-get install -yq --no-install-recommends \
   bash-completion      \
   build-essential      \
   git                  \
-  g++-8                \
-  gcc-8                \
+  g++                \
+  gcc                \
   man                  \
   make                 \
   ccache               \
@@ -20,19 +20,17 @@ apt-get clean
 pushd $INSTALL_DIR
 
 # CMake
-curl -o cmake.tgz -L https://github.com/Kitware/CMake/releases/download/v3.13.4/cmake-3.13.4-Linux-x86_64.tar.gz
+curl -o cmake.tgz -L https://github.com/Kitware/CMake/releases/download/v3.17.5/cmake-3.17.5-Linux-x86_64.tar.gz
 tar xzf cmake.tgz && rm cmake.tgz
-ln -s cmake-*x86_64 cmake
+ln -s cmake-*x86_64 cmake || true
 export PATH="$PWD/cmake/bin:$PATH"
 
 # ccache
 mkdir -p bin
 pushd bin
 # ccache compiler override
-ln -s "$(which ccache)" g++
-ln -s "$(which ccache)" g++-8
-ln -s "$(which ccache)" gcc
-ln -s "$(which ccache)" gcc-8
+ln -s "$(which ccache)" g++ || true
+ln -s "$(which ccache)" gcc || true
 export PATH="$PWD:$PATH"
 popd