+++ docker/user_python-ubuntu.sh
@@ -1,8 +1,27 @@
 #!/usr/bin/env bash
+
+# User can pass 2 parameters to this script: 
+# 1. PYTHON_EXECUTABLE which can be either the 'miniconda' string or the location of the Python executable
+# 2. PYTHON_INSTALL_DIR which points to the location where the user has their virtual environment installed
+#    The virtual environment can be either virtualenv or conda
+#
+# To pass these parameters the user needs to set (at least) temporary environment variables:
+# PYTHON_EXECUTABLE=python3 PYTHON_INSTALL_DIR="~/virtualenv" bash user_python-ubuntu.sh
+# For details see https://github.com/SyneRBI/SIRF-SuperBuild/pull/692#issuecomment-1102704682 
 [ -f .bashrc ] && . .bashrc
-set -ev
-INSTALL_DIR="${1:-/opt/pyvenv}"
-PYTHON="${2:-miniconda}"
+set -exv
+
+if [ -n "${PYTHON_EXECUTABLE}" ]; then 
+  PYTHON=$PYTHON_EXECUTABLE
+else
+  PYTHON='miniconda'
+fi
+
+if [[ -n "${PYTHON_INSTALL_DIR}" ]]; then
+  INSTALL_DIR=${PYTHON_INSTALL_DIR}
+else
+  INSTALL_DIR="/opt/conda"
+fi
 
 # Python
 case "$PYTHON" in
@@ -19,7 +38,7 @@ miniconda)
   ;;
 *python*)
   # virtualenv
-  curl https://bootstrap.pypa.io/get-pip.py > get-pip.py
+  curl $($PYTHON get_pip_download_link.py) > get-pip.py
   ${PYTHON} get-pip.py
   rm get-pip.py
   ${PYTHON} -m pip install -U pip virtualenv
@@ -32,15 +51,15 @@ miniconda)
   ;;
 esac
 
-# Python (runtime)
-if [ -f requirements.txt ]; then
-  pip install -U -r requirements.txt
-fi
-
 if [ "$PYTHON" = "miniconda" ]; then
   if [ -f requirements_conda_forge.txt ]; then
-    conda install --yes -c conda-forge --file requirements_conda_forge.txt
+    conda install --yes -c conda-forge -c intel -c ccpi -c astra-toolbox --file requirements_conda_forge.txt
   fi
   conda update -c conda-forge -y --all
   conda clean -y --all
+fi
+
+# Python (runtime)
+if [ -f requirements.txt ]; then
+  pip install -U -r requirements.txt
 fi